requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.33.0
django-jet==1.0.6
django-tinymce==2.6.0
idna==2.6
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
